Close gosec findings on bounds and file modes
Compare parsed array indices in uint64 space instead of narrowing them
first, bound the predicate index against the slice it came from, keep an
oversized uint64 out of int, read the Vault token from the resolved home
directory, drop the always-taken branch around the path separator, and
tighten the directory and file modes the CLI and tests create.

Credential-shaped literals and the built-binary exec in tests are
fixture data, excluded by path.

+++ pkg/graft/tree/glob.go
@@ -69,7 +69,7 @@ func (c *Cursor) Glob(tree interface{}) ([]*Cursor, error) {
 				i, err := strconv.ParseUint(k, 10, 0)
 				if err == nil {
 					// if k is an integer (in string form), go by index
-					if int(i) >= len(val) {
+					if i >= uint64(len(val)) {
 						return nil, NotFoundError{
 							Path: path[0 : pos+1],
 						}
